IPify added, code reorganized
IPify added to help find public IP of local machine, GetLocationExample
merged with IPIfy code to form IPLocationEstimator, obsolete class
removed. The IP location component seems to be working fully.
