+++ .classpath
@@ -6,5 +6,6 @@
 	<classpathentry kind="lib" path="libs/JSoup/jsoup-1.9.2.jar"/>
 	<classpathentry kind="lib" path="libs/SQLite/sqlite-jdbc-3.14.2.1.jar"/>
 	<classpathentry kind="lib" path="libs/Maxmind/geoip-api-1.3.1.jar"/>
+	<classpathentry kind="lib" path="libs/IPify/JavaIpify.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
